refactor: use camel case naming for pre release variables
