+++ src/lib/is-pre-release.js
