+++ test/is-pre-release.test.js
@@ -1,11 +1,11 @@
 import test from 'ava';
-import getPreRelease from '../src/lib/get-pre-release';
+import isPreRelease from '../src/lib/is-pre-release';
 
 test('No pkg release option', t => {
   const pkg = {};
   const npm = {tag: 'latest'};
 
-  const preRelease = getPreRelease({pkg, npm});
+  const preRelease = isPreRelease({pkg, npm});
 
   t.false(preRelease);
 });
@@ -14,7 +14,7 @@ test('No release.no-pre-release option', t => {
   const pkg = {release: {}};
   const npm = {tag: 'latest'};
 
-  const preRelease = getPreRelease({pkg, npm});
+  const preRelease = isPreRelease({pkg, npm});
 
   t.false(preRelease);
 });
@@ -23,7 +23,7 @@ test('Invalid no-pre-release option', t => {
   const pkg = {release: {'no-pre-release': {}}};
   const npm = {tag: 'xpto2'};
 
-  const preRelease = getPreRelease({pkg, npm});
+  const preRelease = isPreRelease({pkg, npm});
 
   t.false(preRelease);
 });
@@ -32,7 +32,7 @@ test('Empty no-pre-release option', t => {
   const pkg = {release: {'no-pre-release': []}};
   const npm = {tag: 'latest'};
 
-  const preRelease = getPreRelease({pkg, npm});
+  const preRelease = isPreRelease({pkg, npm});
 
   t.false(preRelease);
 });
@@ -42,7 +42,7 @@ test('Dist tag is equal to no-pre-release string option', t => {
   const pkg = {release: {'no-pre-release': distTag}};
   const npm = {tag: distTag};
 
-  const preRelease = getPreRelease({pkg, npm});
+  const preRelease = isPreRelease({pkg, npm});
 
   t.false(preRelease);
 });
@@ -52,7 +52,7 @@ test('Dist Tag is included in no-pre-release single value array option', t => {
   const pkg = {release: {'no-pre-release': [distTag]}};
   const npm = {tag: distTag};
 
-  const preRelease = getPreRelease({pkg, npm});
+  const preRelease = isPreRelease({pkg, npm});
 
   t.false(preRelease);
 });
@@ -62,7 +62,7 @@ test('Dist tag is included in no-pre-release multiple value array option', t => 
   const pkg = {release: {'no-pre-release': [distTag, 'yetAnotherValue']}};
   const npm = {tag: distTag};
 
-  const preRelease = getPreRelease({pkg, npm});
+  const preRelease = isPreRelease({pkg, npm});
 
   t.false(preRelease);
 });
@@ -71,7 +71,7 @@ test('Distinct string no-pre-release option', t => {
   const pkg = {release: {'no-pre-release': 'xpto'}};
   const npm = {tag: 'xpto2'};
 
-  const preRelease = getPreRelease({pkg, npm});
+  const preRelease = isPreRelease({pkg, npm});
 
   t.true(preRelease);
 });
@@ -80,7 +80,7 @@ test('Distinct array no-pre-release option', t => {
   const pkg = {release: {'no-pre-release': ['xpto']}};
   const npm = {tag: 'xpto2'};
 
-  const preRelease = getPreRelease({pkg, npm});
+  const preRelease = isPreRelease({pkg, npm});
 
   t.true(preRelease);
 });
@@ -89,7 +89,7 @@ test('Distinct multiple value array no-pre-release option', t => {
   const pkg = {release: {'no-pre-release': ['xpto', 'yetAnotherValue']}};
   const npm = {tag: 'xpto2'};
 
-  const preRelease = getPreRelease({pkg, npm});
+  const preRelease = isPreRelease({pkg, npm});
 
   t.true(preRelease);
 });
@@ -98,7 +98,7 @@ test('Dist tag is not latest', t => {
   const pkg = {release: {}};
   const npm = {tag: 'xpto2'};
 
-  const preRelease = getPreRelease({pkg, npm});
+  const preRelease = isPreRelease({pkg, npm});
 
   t.true(preRelease);
 });
